refactor: Move the settings test fixtures out of the test file
The file was over the size the linter allows, with five hundred
lines of fixtures ahead of the tests.
