+++ ios/Packages/Settings/Tests/SettingsTests/SettingsScreenTest.swift
@@ -7,498 +7,6 @@ import SwiftUI
 import XCTest
 
 class SettingsScreenTest: SnapshotTestCase {
-    private let sampleThemes: [ThemeItemModel] = [
-        ThemeItemModel(
-            id: "system",
-            displayName: "System",
-            backgroundColor: TvManiacColorScheme.light.background,
-            accentColor: TvManiacColorScheme.light.secondary,
-            onAccentColor: TvManiacColorScheme.light.onSecondary,
-            isSystemTheme: true
-        ),
-        ThemeItemModel(
-            id: "light",
-            displayName: "Light",
-            backgroundColor: TvManiacColorScheme.light.background,
-            accentColor: TvManiacColorScheme.light.secondary,
-            onAccentColor: TvManiacColorScheme.light.onSecondary
-        ),
-        ThemeItemModel(
-            id: "dark",
-            displayName: "Dark",
-            backgroundColor: TvManiacColorScheme.dark.background,
-            accentColor: TvManiacColorScheme.dark.secondary,
-            onAccentColor: TvManiacColorScheme.dark.onSecondary
-        ),
-        ThemeItemModel(
-            id: "autumn",
-            displayName: "Autumn",
-            backgroundColor: TvManiacColorScheme.autumn.background,
-            accentColor: TvManiacColorScheme.autumn.secondary,
-            onAccentColor: TvManiacColorScheme.autumn.onSecondary,
-            isPremium: true
-        ),
-        ThemeItemModel(
-            id: "aqua",
-            displayName: "Aqua",
-            backgroundColor: TvManiacColorScheme.aqua.background,
-            accentColor: TvManiacColorScheme.aqua.secondary,
-            onAccentColor: TvManiacColorScheme.aqua.onSecondary,
-            isPremium: true
-        ),
-        ThemeItemModel(
-            id: "amber",
-            displayName: "Amber",
-            backgroundColor: TvManiacColorScheme.amber.background,
-            accentColor: TvManiacColorScheme.amber.secondary,
-            onAccentColor: TvManiacColorScheme.amber.onSecondary,
-            isPremium: true
-        ),
-        ThemeItemModel(
-            id: "snow",
-            displayName: "Snow",
-            backgroundColor: TvManiacColorScheme.snow.background,
-            accentColor: TvManiacColorScheme.snow.secondary,
-            onAccentColor: TvManiacColorScheme.snow.onSecondary,
-            isPremium: true
-        ),
-        ThemeItemModel(
-            id: "terminal",
-            displayName: "Terminal",
-            backgroundColor: TvManiacColorScheme.terminal.background,
-            accentColor: TvManiacColorScheme.terminal.secondary,
-            onAccentColor: TvManiacColorScheme.terminal.onSecondary,
-            isPremium: true
-        ),
-        ThemeItemModel(
-            id: "crimson",
-            displayName: "Crimson",
-            backgroundColor: TvManiacColorScheme.crimson.background,
-            accentColor: TvManiacColorScheme.crimson.secondary,
-            onAccentColor: TvManiacColorScheme.crimson.onSecondary,
-            isPremium: true
-        ),
-    ]
-
-    private var defaultThemeItem: SettingsThemeItem<ThemeItemModel> {
-        SettingsThemeItem(
-            icon: "paintpalette",
-            title: "Theme",
-            subtitle: "Choose your preferred theme",
-            themes: sampleThemes,
-            selectedTheme: sampleThemes[0],
-            onThemeSelected: { _ in }
-        )
-    }
-
-    private var customThemeItem: SettingsThemeItem<ThemeItemModel> {
-        SettingsThemeItem(
-            icon: "paintpalette",
-            title: "Theme",
-            subtitle: "Choose your preferred theme",
-            themes: sampleThemes,
-            selectedTheme: sampleThemes[0],
-            isCustomThemesLocked: true,
-            lockedBadgeText: "Premium",
-            lockedTitle: "Custom themes are a Premium feature",
-            lockedMessage: "Upgrade to Premium to use custom themes.",
-            lockedActionText: "Upgrade to Premium",
-            lockedAccessibilityLabel: "Locked",
-            onThemeSelected: { _ in }
-        )
-    }
-
-    private var defaultImageQualityItem: SettingsImageQualityItem {
-        SettingsImageQualityItem(
-            icon: "photo",
-            title: "Image Quality",
-            subtitle: "Automatically adjusts based on network",
-            options: [
-                SettingsImageQualityOption(id: "AUTO", label: "Auto", onSelect: {}),
-                SettingsImageQualityOption(id: "HIGH", label: "High", onSelect: {}),
-                SettingsImageQualityOption(id: "MEDIUM", label: "Medium", onSelect: {}),
-                SettingsImageQualityOption(id: "LOW", label: "Low", onSelect: {}),
-            ],
-            selectedOptionId: "AUTO"
-        )
-    }
-
-    private var layoutToggles: [SettingsToggleItem] {
-        [
-            SettingsToggleItem(
-                id: "haptic",
-                icon: "iphone.radiowaves.left.and.right",
-                title: "Haptic feedback",
-                subtitle: "Feel subtle vibrations during interactions",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "season-order",
-                icon: "arrow.up.arrow.down",
-                title: "Season Order",
-                subtitle: "Order the latest season first",
-                isOn: false,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "blur-unwatched",
-                icon: "eye.slash",
-                title: "Hide Spoilers",
-                subtitle: "Hide spoilers for unwatched episodes",
-                isOn: false,
-                onToggle: { _ in }
-            ),
-        ]
-    }
-
-    private func fontSizeItem(percent: Int = 100) -> SettingsFontSizeItem {
-        SettingsFontSizeItem(
-            title: "Font Size",
-            description: "Adjust text size across the app",
-            previewText: "The quick brown fox jumps over the lazy dog",
-            resetLabel: "Reset",
-            percent: percent,
-            onPercentChange: { _ in }
-        )
-    }
-
-    private var discoverSectionsNavItem: SettingsNavigationItem {
-        SettingsNavigationItem(
-            id: SettingsPageRoute.discoverSections.rawValue,
-            icon: SettingsPageRoute.discoverSections.iconName,
-            title: "Discover Sections",
-            subtitle: "Choose which sections appear on Discover",
-            onTap: {}
-        )
-    }
-
-    private var discoverSectionToggles: [SettingsToggleItem] {
-        [
-            SettingsToggleItem(
-                id: "START_WATCHING",
-                icon: "play.circle",
-                title: "Start Watching",
-                subtitle: "",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "TRENDING_TODAY",
-                icon: "flame",
-                title: "Trending Today",
-                subtitle: "",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "UPCOMING",
-                icon: "calendar",
-                title: "Upcoming",
-                subtitle: "",
-                isOn: false,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "POPULAR",
-                icon: "star",
-                title: "Popular",
-                subtitle: "",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "TOP_RATED",
-                icon: "trophy",
-                title: "Top Rated",
-                subtitle: "",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-        ]
-    }
-
-    private func posterStyleItem(locked: Bool = false) -> SettingsPosterStyleItem {
-        let widthOptions: [SettingsPosterStyleOption] = [
-            SettingsPosterStyleOption(id: "COMPACT", label: "Compact", onSelect: {}),
-            SettingsPosterStyleOption(id: "STANDARD", label: "Standard", onSelect: {}),
-            SettingsPosterStyleOption(id: "COMFORTABLE", label: "Comfortable", onSelect: {}),
-            SettingsPosterStyleOption(id: "LARGE", label: "Large", onSelect: {}),
-        ]
-        let cornerOptions: [SettingsPosterStyleOption] = [
-            SettingsPosterStyleOption(id: "SHARP", label: "Sharp", onSelect: {}),
-            SettingsPosterStyleOption(id: "CLASSIC", label: "Classic", onSelect: {}),
-            SettingsPosterStyleOption(id: "ROUNDED", label: "Rounded", onSelect: {}),
-            SettingsPosterStyleOption(id: "PILL", label: "Pill", onSelect: {}),
-        ]
-        return SettingsPosterStyleItem(
-            title: "Poster style",
-            description: "Choose poster size and corner style",
-            livePreviewLabel: "Live preview",
-            resetLabel: "Reset",
-            postersLabel: "Posters",
-            landscapeLabel: "Landscape",
-            cornerLabel: "Corner style",
-            postersOptions: widthOptions,
-            landscapeOptions: widthOptions,
-            cornerOptions: cornerOptions,
-            selectedPostersId: "STANDARD",
-            selectedLandscapeId: "STANDARD",
-            selectedCornerId: "SHARP",
-            posterScale: 1,
-            landscapeScale: 1,
-            cornerRadius: 0,
-            isLocked: locked,
-            lockedBadgeText: locked ? "Premium" : "",
-            lockedActionText: locked ? "Upgrade to Premium" : "",
-            lockedAccessibilityLabel: locked ? "Locked" : ""
-        )
-    }
-
-    private var posterStyleNavItem: SettingsNavigationItem {
-        SettingsNavigationItem(
-            id: SettingsPageRoute.posterStyle.rawValue,
-            icon: SettingsPageRoute.posterStyle.iconName,
-            title: "Poster style",
-            subtitle: "Choose poster size and corner style",
-            onTap: {}
-        )
-    }
-
-    private var behaviorToggles: [SettingsToggleItem] {
-        [
-            SettingsToggleItem(
-                id: "sync",
-                icon: "arrow.triangle.2.circlepath",
-                title: "Background Sync",
-                subtitle: "Sync your library in the background",
-                isOn: false,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "specials",
-                icon: "film.stack",
-                title: "Include Specials",
-                subtitle: "Show special episodes in season lists",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-            SettingsToggleItem(
-                id: "youtube",
-                icon: "tv",
-                title: "Open in YouTube",
-                subtitle: "Open trailers in the YouTube app",
-                isOn: false,
-                onToggle: { _ in }
-            ),
-        ]
-    }
-
-    private var notificationToggles: [SettingsToggleItem] {
-        [
-            SettingsToggleItem(
-                id: "notifications",
-                icon: "bell.fill",
-                title: "Episode Notifications",
-                subtitle: "Get notified when new episodes air",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-        ]
-    }
-
-    private var lockedNotificationToggles: [SettingsToggleItem] {
-        [
-            SettingsToggleItem(
-                id: "notifications",
-                icon: "bell.fill",
-                title: "Episode Notifications",
-                subtitle: "Get notified when new episodes air",
-                isOn: false,
-                isLocked: true,
-                lockedBadgeText: "Premium",
-                lockedAccessibilityLabel: "Locked",
-                onToggle: { _ in }
-            ),
-        ]
-    }
-
-    private var privacyToggles: [SettingsToggleItem] {
-        [
-            SettingsToggleItem(
-                id: "crash-reporting",
-                icon: "ladybug",
-                title: "Crash Reporting",
-                subtitle: "Help improve the app by sending crash reports",
-                isOn: true,
-                onToggle: { _ in }
-            ),
-        ]
-    }
-
-    private var privacyLinks: [SettingsNavigationItem] {
-        [
-            SettingsNavigationItem(
-                id: "privacy-policy",
-                icon: "hand.raised",
-                title: "Privacy Policy",
-                onTap: {}
-            ),
-        ]
-    }
-
-    private var infoContent: SettingsInfoContent {
-        SettingsInfoContent(
-            icon: Image(systemName: "app.fill"),
-            appName: "TvManiac",
-            versionText: "Version 1.0.0",
-            description: "A Kotlin Multiplatform app for discovering and tracking your favorite TV shows.",
-            sourceCodeLabel: "Source Code",
-            sourceCodeValue: "GitHub",
-            apiDisclaimer: "This product uses the TMDB and Trakt API but is not endorsed or certified by either.",
-            onVersionTap: {},
-            onSourceCodeTap: {}
-        )
-    }
-
-    private var licenseSections: [SettingsLicenseSection] {
-        [
-            SettingsLicenseSection(id: "app", label: "App", items: [
-                SettingsLinkItem(
-                    id: "tvmaniac",
-                    title: "TvManiac",
-                    body: "Open-source on GitHub.",
-                    link: "https://github.com/c0de-wizard/tv-maniac",
-                    onOpen: {}
-                ),
-            ]),
-            SettingsLicenseSection(id: "data", label: "Data & Services", items: [
-                SettingsLinkItem(
-                    id: "tmdb",
-                    leadingAsset: "TmdbLogo",
-                    title: "The Movie Database (TMDB)",
-                    body: "This product uses the TMDB API but is not endorsed or certified by TMDB.",
-                    link: "https://www.themoviedb.org",
-                    onOpen: {}
-                ),
-                SettingsLinkItem(
-                    id: "trakt",
-                    leadingAsset: "TraktLogo",
-                    title: "Trakt",
-                    body: "Syncs your watch history, watchlist, and episode progress.",
-                    link: "https://trakt.tv",
-                    onOpen: {}
-                ),
-            ]),
-        ]
-    }
-
-    private func accountContent(
-        authenticated: Bool,
-        withSwitchAffordance: Bool = false,
-        isSwitching: Bool = false,
-        showSwitchConfirmation: Bool = false,
-        isProcessingAuth: Bool = false
-    ) -> SettingsAccountContent {
-        let switchLabel: String? = withSwitchAffordance || isSwitching || showSwitchConfirmation
-            ? "Switch to Simkl"
-            : nil
-        return SettingsAccountContent(
-            title: "Trakt",
-            description: "Sync your watchlist, watch progress, continue watching, and personal lists with Trakt.",
-            authenticationLabel: "Connect & Sync Your Content",
-            connectTitle: "Connect",
-            syncDescription: "Save your progress, discover new titles, and sync your content across all devices.",
-            connectedTitle: authenticated ? "Connected" : "Connect to Trakt",
-            connectedDescription: authenticated
-                ? "Your watch history, watchlist, and episode progress sync with Trakt."
-                : "Sign in with Trakt to sync your watch history, watchlist, and episode progress across your devices.",
-            isAuthenticated: authenticated,
-            isProcessingAuth: isProcessingAuth,
-            logoutLabel: "Logout",
-            loginLabel: "Login",
-            providerName: "Trakt",
-            authProviders: [
-                SwiftAuthProvider(id: "TRAKT", label: "Continue with Trakt", logoName: "TraktMono"),
-                SwiftAuthProvider(id: "SIMKL", label: "Continue with Simkl", logoName: "SimklMono"),
-            ],
-            switchTargetLogoName: switchLabel != nil ? "SimklMono" : nil,
-            switchActionLabel: switchLabel,
-            isSwitching: isSwitching,
-            showSwitchConfirmation: showSwitchConfirmation,
-            switchDialogTitle: showSwitchConfirmation ? "Switch to Simkl?" : nil,
-            switchDialogMessage: showSwitchConfirmation
-                ? "You have 3 unsaved changes. Switching providers will discard them."
-                : nil,
-            switchConfirmLabel: "Switch",
-            switchCancelLabel: "Cancel",
-            switchingLabel: "Switching...",
-            onLogout: {},
-            onProviderSelected: { _ in },
-            onSwitchProvider: {},
-            onConfirmSwitch: {},
-            onDismissSwitchDialog: {}
-        )
-    }
-
-    private func rootSections(authenticated: Bool) -> [SettingsRootSection] {
-        var sections: [SettingsRootSection] = []
-        if authenticated {
-            sections.append(SettingsRootSection(id: "account", label: "Account", items: [
-                navItem(.account, "Trakt Account", "Manage your Trakt connection"),
-            ]))
-        }
-        sections.append(SettingsRootSection(id: "general", label: "General", items: [
-            navItem(.appearance, "Appearance", "Theme and image quality"),
-            navItem(.layout, "Layout", "Personalize how shows and episodes look"),
-            navItem(.behavior, "Behavior", "Sync, specials, and trailers"),
-            navItem(.notifications, "Notifications", "Episode release alerts"),
-            navItem(.privacy, "Privacy", "Crash reporting and privacy policy"),
-        ]))
-        sections.append(SettingsRootSection(id: "about", label: "About", items: [
-            navItem(.info, "Info", "App version and source code"),
-            navItem(.licenses, "Licenses & Attribution", "Data sources and acknowledgements"),
-        ]))
-        return sections
-    }
-
-    private func navItem(_ route: SettingsPageRoute, _ title: String, _ subtitle: String) -> SettingsNavigationItem {
-        SettingsNavigationItem(id: route.rawValue, icon: route.iconName, title: title, subtitle: subtitle, onTap: {})
-    }
-
-    private func makeState(
-        page: SettingsPageRoute,
-        authenticated: Bool,
-        isLoading: Bool = false,
-        customAccountContent: SettingsAccountContent? = nil,
-        customThemeItem: SettingsThemeItem<ThemeItemModel>? = nil,
-        customNotificationToggles: [SettingsToggleItem]? = nil,
-        fontSizePercent: Int = 100,
-        customPosterStyleItem: SettingsPosterStyleItem? = nil
-    ) -> SettingsScreen<ThemeItemModel>.State {
-        SettingsScreen<ThemeItemModel>.State(
-            isLoading: isLoading,
-            rootTitle: "Settings",
-            currentPage: page,
-            rootSections: rootSections(authenticated: authenticated),
-            themeItem: customThemeItem ?? defaultThemeItem,
-            imageQualityItem: defaultImageQualityItem,
-            layoutToggles: layoutToggles,
-            fontSizeItem: fontSizeItem(percent: fontSizePercent),
-            discoverSectionsNavItem: discoverSectionsNavItem,
-            discoverSectionToggles: discoverSectionToggles,
-            posterStyleNavItem: posterStyleNavItem,
-            posterStyleItem: customPosterStyleItem ?? posterStyleItem(),
-            behaviorToggles: behaviorToggles,
-            notificationToggles: customNotificationToggles ?? notificationToggles,
-            privacyToggles: privacyToggles,
-            privacyLinks: privacyLinks,
-            infoContent: infoContent,
-            licenseSections: licenseSections,
-            accountContent: customAccountContent ?? accountContent(authenticated: authenticated)
-        )
-    }
-
     func test_SettingsScreen_Loading() {
         SettingsScreen(state: makeState(page: .root, authenticated: false, isLoading: true), onBack: {})
             .appPreview()
